requirements.txt
discord.py
aiohttp
deepl
googletrans==4.0.0-rc1
python-dotenv